[ovsp4rt_v3] Move encode utilities to ovsp4rt_decode_utils.cc
