+++ ovs-p4rt/sidecar/core/v3/CMakeLists.txt
@@ -9,6 +9,7 @@ if(DPDK_TARGET)
         dpdk_config_fdb_entry.cc
         dpdk_config_tunnel_entry.cc
         dpdk_encoders.cc
+        ovsp4rt_encode_utils.cc
         ovsp4rt_p4info_utils.cc
     )
 elseif(ES2K_TARGET)
@@ -21,6 +22,7 @@ elseif(ES2K_TARGET)
         es2k_config_fdb_entry.cc
         es2k_config_tunnel_entry.cc
         es2k_encoders.cc
+        ovsp4rt_encode_utils.cc
         ovsp4rt_p4info_utils.cc
         tx_acc_vsi_table_entry.cc
         tx_acc_vsi_table_entry.h
