+++ ovs-p4rt/sidecar/core/v3/dpdk_encoders.cc
@@ -18,43 +18,6 @@
 
 namespace ovsp4rt {
 
-std::string EncodeByteValue(int arg_count...) {
-  std::string byte_value;
-  va_list args;
-  va_start(args, arg_count);
-
-  for (int arg = 0; arg < arg_count; ++arg) {
-    uint8_t byte = va_arg(args, int);
-    byte_value.push_back(byte);
-  }
-
-  va_end(args);
-  return byte_value;
-}
-
-// Encodes tunnel_info.vni as a "tunnel_id" action parameter,
-// which is bit<20> in all cases except set_ipsec_tunnel.
-static inline std::string EncodeTunnelId(uint32_t vni) {
-  return EncodeByteValue(3, (vni >> 16) & 0x0F, (vni >> 8) & 0xFF, vni & 0xFF);
-}
-
-// Encodes tunnel_info.vni as a "vni" or "mod_blob_ptr" match
-// field or action parameter, which are bit<24> in all cases.
-static inline std::string EncodeVniValue(uint32_t vni) {
-  return EncodeByteValue(3, (vni >> 16) & 0xFF, (vni >> 8) & 0xFF, vni & 0xFF);
-}
-
-std::string CanonicalizeIp(const uint32_t ipv4addr) {
-  // note: low-to-high byte order
-  return EncodeByteValue(4, (ipv4addr & 0xff), ((ipv4addr >> 8) & 0xff),
-                         ((ipv4addr >> 16) & 0xff), ((ipv4addr >> 24) & 0xff));
-}
-
-std::string CanonicalizeMac(const uint8_t mac[6]) {
-  return EncodeByteValue(6, (mac[0] & 0xff), (mac[1] & 0xff), (mac[2] & 0xff),
-                         (mac[3] & 0xff), (mac[4] & 0xff), (mac[5] & 0xff));
-}
-
 void EncodeL2FwdTxTablePrologue(p4::v1::TableEntry* table_entry,
                                 const struct mac_learning_info& learn_info,
                                 const ::p4::config::v1::P4Info& p4info) {
